.claude/worktrees/blissful-swirles/brain/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
langchain>=0.1.0
langchain-google-genai>=0.1.0
google-generativeai>=0.7.0
pydantic>=2.0.0
tiktoken>=0.5.0
